requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.4.1
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
